Add get_utxo_proof method
