+++ api/src/method/mod.rs
@@ -2,5 +2,6 @@ pub mod get_compressed_account;
 pub mod get_compressed_account_proof;
 pub mod get_compressed_token_accounts_by_owner;
 pub mod get_utxo;
+pub mod get_utxo_proof;
 pub mod get_utxos;
 pub mod utils;
